# storage-service/docs/swagger.yaml
basePath: /api/storage
definitions:
  storage-service_internal_models.Object:
    description: 3D object information
    properties:
      content_type:
        type: string
      id:
        type: string
      original_filename:
        type: string
      size:
        type: integer
      storage_key:
        type: string
      uploaded_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Storage Service API for managing 3D objects
  title: Storage Service API
  version: "1.0"
paths:
  /objects:
    get:
      consumes:
      - application/json
      description: Gets all 3D objects stored in the system
      produces:
      - application/json
      responses:
        "200":
          description: List of all 3D objects
          schema:
            items:
              $ref: '#/definitions/storage-service_internal_models.Object'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: List all 3D objects
      tags:
      - objects
    post:
      consumes:
      - multipart/form-data
      description: 'Upload one or more 3D model files (formats: .fbx, .obj, .dae,
        .stl, .gltf). Multiple files can be provided for a model with external resources
        (e.g. textures).'
      parameters:
      - description: 3D model file(s)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Object successfully created
          schema:
            $ref: '#/definitions/storage-service_internal_models.Object'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Upload a new 3D object
      tags:
      - objects
  /objects/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a 3D object by ID
      parameters:
      - description: Object ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid UUID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Object not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a 3D object
      tags:
      - objects
    get:
      consumes:
      - application/json
      description: Get details of a specific 3D object
      parameters:
      - description: Object ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Object Found
          schema:
            $ref: '#/definitions/storage-service_internal_models.Object'
        "400":
          description: Invalid UUID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Object not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get a 3D object by ID
      tags:
      - objects
    put:
      consumes:
      - multipart/form-data
      description: Replace an existing 3D object file
      parameters:
      - description: Object ID
        in: path
        name: id
        required: true
        type: string
      - description: 3D object file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Object not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update a 3D object
      tags:
      - objects
  /objects/{id}/download:
    get:
      consumes:
      - application/json
      description: Download the GLB file for a specific 3D object
      parameters:
      - description: Object ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: GLB file
          schema:
            type: file
        "400":
          description: Invalid UUID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Object not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Download a 3D object file
      tags:
      - objects
swagger: "2.0"